fix: live 와 live-session 프로젝트  통합
