+++ src/main/java/com/developers/live/config/CorsConfig.java
@@ -1,0 +1,28 @@
+package com.developers.live.config;
+
+import org.springframework.context.annotation.Bean;
+import org.springframework.context.annotation.Configuration;
+import org.springframework.web.cors.CorsConfiguration;
+import org.springframework.web.cors.UrlBasedCorsConfigurationSource;
+import org.springframework.web.filter.CorsFilter;
+
+import java.util.Arrays;
+
+@Configuration
+public class CorsConfig {
+
+  @Bean
+  public CorsFilter corsFilter() {
+    UrlBasedCorsConfigurationSource source = new UrlBasedCorsConfigurationSource();
+    CorsConfiguration config = new CorsConfiguration();
+
+    config.setAllowCredentials(true);
+    config.addAllowedOriginPattern("http://localhost:3000");
+    config.setAllowedMethods(Arrays.asList("GET","POST","DELETE","UPDATE","OPTIONS"));
+    config.setAllowedHeaders(Arrays.asList("Origin", "Accept", "X-Requested-With", "Content-Type", "Access-Control-Request-Method", "Access-Control-Request-Headers"));
+    config.addExposedHeader("*");
+    source.registerCorsConfiguration("/**", config);
+
+    return new CorsFilter(source);
+  }
+}
