+++ src/main/java/com/developers/live/session/dto/SessionRedisFindAllResponse.java
@@ -1,0 +1,16 @@
+package com.developers.live.session.dto;
+
+import lombok.AllArgsConstructor;
+import lombok.Builder;
+import lombok.Data;
+import lombok.NoArgsConstructor;
+
+@AllArgsConstructor
+@NoArgsConstructor
+@Builder
+@Data
+public class SessionRedisFindAllResponse {
+    private String code;
+    private String msg;
+    private String data;
+}
