+++ src/main/java/com/developers/live/session/dto/SessionRedisRemoveRequest.java
@@ -1,0 +1,16 @@
+package com.developers.live.session.dto;
+
+import jakarta.validation.constraints.NotNull;
+import lombok.AllArgsConstructor;
+import lombok.Builder;
+import lombok.Data;
+import lombok.NoArgsConstructor;
+
+@AllArgsConstructor
+@NoArgsConstructor
+@Builder
+@Data
+public class SessionRedisRemoveRequest {
+    @NotNull
+    private String roomName;
+}
