+++ src/main/java/com/developers/live/session/dto/SessionRedisSaveRequest.java
@@ -1,0 +1,19 @@
+package com.developers.live.session.dto;
+
+import jakarta.validation.constraints.NotNull;
+import lombok.AllArgsConstructor;
+import lombok.Data;
+import lombok.NoArgsConstructor;
+
+@AllArgsConstructor
+@NoArgsConstructor
+@Data
+public class SessionRedisSaveRequest {
+    @NotNull
+    private String roomName;
+    @NotNull
+    private String userName;
+
+    @NotNull
+    private Long time;
+}
