+++ src/main/java/com/developers/live/session/dto/SessionRedisSaveResponse.java
@@ -1,0 +1,17 @@
+package com.developers.live.session.dto;
+
+
+import lombok.AllArgsConstructor;
+import lombok.Builder;
+import lombok.Data;
+import lombok.NoArgsConstructor;
+
+@AllArgsConstructor
+@NoArgsConstructor
+@Builder
+@Data
+public class SessionRedisSaveResponse {
+    private String code;
+    private String msg;
+    private String data;
+}
