+++ src/main/java/com/developers/live/session/dto/dailyco/DailyCoAnswer.java
@@ -1,0 +1,18 @@
+package com.developers.live.session.dto.dailyco;
+
+import lombok.AllArgsConstructor;
+import lombok.Getter;
+import lombok.NoArgsConstructor;
+
+@NoArgsConstructor
+@AllArgsConstructor
+@Getter
+public class DailyCoAnswer {
+    private String id;
+    private String name;
+    private boolean api_created;
+    private String privacy;
+    private String url;
+    private String created_at;
+    private Object config;
+}
