+++ src/main/java/com/developers/live/session/dto/dailyco/DailyCoResponse.java
@@ -1,0 +1,16 @@
+package com.developers.live.session.dto.dailyco;
+
+import lombok.AllArgsConstructor;
+import lombok.Builder;
+import lombok.Data;
+import lombok.NoArgsConstructor;
+
+@NoArgsConstructor
+@AllArgsConstructor
+@Builder
+@Data
+public class DailyCoResponse {
+    private String code;
+    private String msg;
+    private DailyCoAnswer data;
+}
